feat: Serialize player hand and add player hand to ElementalsWar view
This commit adds functionality to the ElementalsWar view to serialize the player object, including the player's hand, and add it to the JSON response returned by the view. The serialized player object is then added to the 'player' data property in the ElementalsWar component, and the player's hand is displayed using a new 'player-hand' div in the template. The 'player-hand' div contains a header with the player's name, a list of the player's hand elements, and a message if the player's hand is empty. The CSS for the 'player-hand' div has been added to make it more visually appealing.