"Add EntityImageView to get entity image url and update ElementalsWar.vue to fetch entity images.
The EntityImageView endpoint has been added to elementals_app/urls.py to retrieve the image URL for a given entity type. The ElementalsWar.vue component has been updated to use the new endpoint to fetch the entity image URL for each group of three or more cards in the player's hand. The function fetchEntityImageURL has been added to ElementalsWarMethods.js to handle the API call and return the image URL."