Title: Add is_computer field to Player model and update ElementalsWarView to include computer player
Description: This commit adds a new Boolean field, is_computer, to the Player model. This field is used to differentiate between human and computer players in the ElementalsWarView. This commit updates the ElementalsWarView to include a computer player. The view now returns a serialized representation of both the human and computer players, using the new is_computer field in the Player model.