Add asynchronous functionality to flipCard and computerMove
This commit adds asynchronous functionality to the flipCard and computerMove methods in ElementalsWarMethods.js to prevent players from flipping more than two cards at a time and to allow the computer move to execute only after the player has completed their move.

Additionally, this commit updates the flipCard method to check if the card can be flipped based on the canFlip flag and the state of the selected card.

Lastly, the canFlip and playerTurn flags were added to the initial state of the ElementalsWar Vue component to help manage the state of the game.